Load and render deviceList from API
Note: this is a very hacky use of useEffect and needs refactored, we should also abstract the loading such that FlatList's `refreshing` and `onRefresh` are updated via props.
